+++ src/nodes/xRaySQ.h
@@ -9,13 +9,8 @@
 #include "data/formFactors.h"
 #include "data/structureFactors.h"
 #include "math/windowFunction.h"
-#include "nodes/gr.h"
 #include "nodes/node.h"
 #include <optional>
-
-// Forward Declarations
-class PartialSet;
-class XRayWeights;
 
 class XRaySQNode : public Node
 {
@@ -23,12 +18,15 @@ class XRaySQNode : public Node
     XRaySQNode(Graph *parentGraph);
     ~XRaySQNode() override = default;
 
+    /*
+     * Definition
+     */
     public:
     std::string_view type() const override;
     std::string_view summary() const override;
 
     /*
-     * Definition
+     * Data
      */
     private:
     // Atomic form factors to use for weighting
@@ -63,48 +61,7 @@ class XRaySQNode : public Node
     StructureFactors::NormalisationType referenceNormalisedTo_{StructureFactors::NoNormalisation};
     // Window function to use when Fourier transforming reference total F(Q) into g(r)
     WindowFunction::Form referenceWindowFunction_{WindowFunction::Form::Lorch0};
-    // Whether to save combined form factor weightings for atomtype pairs
-    bool saveFormFactors_{false};
-    // Save weighted g(r) and G(r)
-    bool saveGR_{false};
-    // Save the reference data and its Fourier transform
-    bool saveReference_{false};
-    // Save representative G(r), obtained from Fourier transform of the calculated F(Q)
-    bool saveRepresentativeGR_{false};
-    // Save weighted partial and total structure factors
-    bool saveSQ_{false};
 
-    public:
-    // Return file and format for reference total F(Q)
-    const Data1DImportFileFormat &referenceFQFileAndFormat();
-
-    /*
-     * Functions
-     */
-    private:
-    // Set up reference data
-    bool setReferenceData();
-
-    public:
-    // Return xRay weights
-    const XRayWeights &weights() const;
-    // Calculate weighted g(r) from supplied unweighted g(r) and Weights
-    bool calculateWeightedGR(const PartialSet &unweightedgr, PartialSet &weightedgr, const XRayWeights &weights,
-                             StructureFactors::NormalisationType normalisation);
-    // Calculate weighted S(Q) from supplied unweighted S(Q) and Weights
-    bool calculateWeightedSQ(const PartialSet &unweightedsq, PartialSet &weightedsq, const XRayWeights &weights,
-                             StructureFactors::NormalisationType normalisation);
-
-    /*
-     * Processing
-     */
-    private:
-    // Run main processing
-    NodeConstants::ProcessResult process() override;
-
-    /*
-     * Getters
-     */
     public:
     // Returns the unweighted SQ
     const PartialSet *unweightedSQ() const;
@@ -112,4 +69,23 @@ class XRaySQNode : public Node
     const PartialSet *unweightedGR() const;
     // Returns the source configuration, belonging to the input SQ node
     const Configuration *sourceConfiguration();
+    // Return xRay weights
+    const XRayWeights &weights() const;
+
+    /*
+     * Processing
+     */
+    private:
+    // Set up reference data
+    bool setUpReferenceData();
+    // Calculate weighted g(r) from supplied unweighted g(r) and Weights
+    bool calculateWeightedGR(const PartialSet &unweightedgr, PartialSet &weightedgr, const XRayWeights &weights,
+                             StructureFactors::NormalisationType normalisation);
+    // Calculate weighted S(Q) from supplied unweighted S(Q) and Weights
+    bool calculateWeightedSQ(const PartialSet &unweightedsq, PartialSet &weightedsq, const XRayWeights &weights,
+                             StructureFactors::NormalisationType normalisation);
+
+    protected:
+    // Run main processing
+    NodeConstants::ProcessResult process() override;
 };
